+++ lib/src/query/cause/actor.dart
@@ -8,15 +8,9 @@ import 'package:delta_trace_db/delta_trace_db.dart';
 /// (ja) データベースの操作をリクエストした者の情報を定義するクラスです。
 class Actor extends CloneableFile {
   static const String className = "Actor";
-  static const String version = "3";
+  static const String version = "5";
   final EnumActorType type;
   final String id;
-
-  // Actorが「何者であるか」を示すビジネス上の役割。
-  final List<String> roles;
-
-  // Actorが「何ができるか」を示す具体的な操作権限。命名規則はリソース:アクション:スコープ。
-  final List<String> permissions;
 
   // コレクション単位の操作に関するパーミッション。
   final Map<String, Permission>? collectionPermissions;
@@ -25,21 +19,10 @@ class Actor extends CloneableFile {
 
   /// * [type] : The actor type. Choose from HUMAN, AI, or SYSTEM.
   /// * [id] : The serial id (user id) of the actor.
-  /// * [roles] : A business role that describes who an actor "is."
-  /// * [permissions] : Specific operational permissions that indicate
-  /// "what an actor can do." The naming convention is resource:action:scope.
-  /// e.g. users:read:all => read all user information.
   /// * [collectionPermissions] : Collection-level permissions that relate only
   /// to database operations. The key is the collection name.
   /// * [context] : other context.
-  Actor(
-    this.type,
-    this.id,
-    this.roles,
-    this.permissions, {
-    this.collectionPermissions,
-    this.context,
-  });
+  Actor(this.type, this.id, {this.collectionPermissions, this.context});
 
   /// (en) Recover this class from the dictionary.
   ///
@@ -62,8 +45,6 @@ class Actor extends CloneableFile {
     return Actor(
       EnumActorType.values.byName(src["type"]),
       src["id"],
-      (src["roles"] as List).cast<String>(),
-      (src["permissions"] as List).cast<String>(),
       collectionPermissions: collectionPermissions,
       context: src["context"] != null
           ? src["context"] as Map<String, dynamic>
@@ -90,8 +71,6 @@ class Actor extends CloneableFile {
       "version": version,
       "type": type.name,
       "id": id,
-      "roles": roles,
-      "permissions": permissions,
       "collectionPermissions": mCollectionPermissions,
       "context": UtilCopy.jsonableDeepCopy(context),
     };
@@ -102,8 +81,6 @@ class Actor extends CloneableFile {
     if (other is Actor) {
       return type == other.type &&
           id == other.id &&
-          const ListEquality<String>().equals(roles, other.roles) &&
-          const ListEquality<String>().equals(permissions, other.permissions) &&
           DeepCollectionEquality().equals(
             collectionPermissions,
             other.collectionPermissions,
@@ -119,8 +96,9 @@ class Actor extends CloneableFile {
     return Object.hashAll([
       type,
       id,
-      UtilObjectHash.calcList(roles),
-      UtilObjectHash.calcList(permissions),
+      collectionPermissions != null
+          ? UtilObjectHash.calcMap(collectionPermissions!)
+          : 0,
       context != null ? UtilObjectHash.calcMap(context!) : 0,
     ]);
   }